[91582] [api] feat: Get book list & book's info

+++ src/main/java/com/group8/library_management/repository/BookRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface BookRepository extends JpaRepository<Book, Long>, JpaSpecificationExecutor<Book> {
+public interface BookRepository extends JpaRepository<Book, Integer>, JpaSpecificationExecutor<Book> {
     @EntityGraph(attributePaths = {"author", "publisher", "genre"})
     Page<Book> findAll(Pageable pageable);
 }
